Revert "[SBI-630] Реализована страница востановления и сброса пароля. (#431)"
This reverts commit 8db1c80d19cf97f5a3c5c77379fb039c9e18972c.
