--- services/trackme-sso/frontend/src/assets/password-reset/PasswordReset.css
@@ -1,95 +1,0 @@
-﻿.reset-container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-    background-color: #f5f5f5;
-    font-family: 'Montserrat', sans-serif;
-}
-
-.reset-box {
-    text-align: center;
-    background: #8338EB;
-    padding: 40px;
-    border-radius: 20px;
-    box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.2);
-    color: white;
-    width: 100%;
-    max-width: 400px;
-}
-
-.reset-title {
-    font-size: 36px;
-    margin-bottom: 20px;
-}
-
-.reset-form {
-    display: flex;
-    flex-direction: column;
-    gap: 15px;
-    margin-bottom: 20px;
-    align-items: center;
-}
-
-.reset-button {
-    background: white;
-    color: #8338EB;
-    font-size: 18px;
-    padding: 15px 30px;
-    border: none;
-    border-radius: 10px;
-    cursor: pointer;
-    font-weight: bold;
-    transition: 0.3s;
-}
-
-.reset-button:hover {
-    background: #e6d7fb;
-}
-
-.reset-button:disabled {
-    opacity: 0.6;
-    cursor: not-allowed;
-}
-
-.reset-success {
-    color: #d4f5d4;
-    margin-bottom: 16px;
-}
-
-.reset-error {
-    color: #ffd6d6;
-    font-size: 14px;
-}
-
-.reset-link {
-    color: white;
-    text-decoration: underline;
-}
-
-@media (max-width: 480px) {
-    .reset-container {
-        background: white;
-    }
-
-    .reset-box {
-        background: white;
-        box-shadow: none;
-    }
-
-    .reset-title {
-        color: #44069A;
-        font-weight: 400;
-        margin-bottom: 23px;
-    }
-
-    .reset-button {
-        background-color: #44069A;
-        color: white;
-        font-size: 13px;
-        font-weight: 700;
-        border-radius: 60px;
-        padding: 10px 20px;
-        min-width: 170px;
-    }
-}
