+++ services/trackme-sso/frontend/src/csrf-axios-interceptor.js
@@ -4,7 +4,6 @@ import axios from 'axios';
 let csrfToken = null;
 let csrfHeaderName = "X-CSRF-TOKEN";
 let basePath = process.env.REACT_APP_BASE_PATH || "";
-
 /**
  * Загружает CSRF-токен (только если он еще не получен)
  */
