+++ services/trackme-sso/frontend/src/services/login-service.js
@@ -1,7 +1,7 @@
 import axios from "../csrf-axios-interceptor";
 
 export class LoginAPI {
-    basePath = process.env.REACT_APP_BASE_PATH || "http://localhost:9000";
+    basePath = process.env.REACT_APP_BASE_PATH || "";
     __LOGIN_URL = "/client/login";
     __LOCATION_HEADER = process.env.REACT_APP_SSO_LOCATION_HEADER;
 
@@ -24,23 +24,6 @@ export class LoginAPI {
             withCredentials: true
         });
     }
-
-    recoveryPassword(email) {
-        return axios.post(`${(this.basePath)}/api/v1/registration/recovery-password`, 
-            { email },
-            { withCredentials: true }
-        );
-    }
-
-    resetPassword(token, password) {
-        return axios.post(`${this.basePath}/api/v1/registration/reset-password`,
-            { password },
-            {
-                params: { token },
-                withCredentials: true
-            }
-    );
-}
 }
 
 let loginAPI = new LoginAPI();
